% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{AA_Prob}
\alias{AA_Prob}
\title{Probability of achieving a specified return}
\usage{
AA_Prob(target, r, sd, t)
}
\arguments{
\item{target}{Target (specified) return}

\item{r}{Mean Return}

\item{sd}{Standard deviation of returns}

\item{t}{Time in years}
}
\value{
value representing a probability.
}
\description{
Returns the probability of doing at least as well as over t
}
\examples{
AA_Prob(7.339443,8,12,10)
}
