% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim-package.R
\docType{package}
\name{aasim-package}
\alias{aasim}
\alias{aasim-package}
\title{aasim: Asset Allocation Simulation}
\description{
Applies Monte Carlo simulation to asset allocation.
}
\author{
\strong{Maintainer}: Rex Macey \email{rex@macey.us}

}
\keyword{internal}
