% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{addPerson.sim}
\alias{addPerson.sim}
\title{Add Person to persons list in a simulation}
\usage{
addPerson.sim(
  sim,
  name,
  initials,
  curAge,
  gender,
  retireAge,
  mort.factor = 1,
  mort.adj.years = 0
)
}
\arguments{
\item{sim}{Object of type sim (simulation)}

\item{name}{Name of person}

\item{initials}{Initials or short name, useful for display}

\item{curAge}{Current age (simulation assumes person just turned this age)}

\item{gender}{'M' or 'Male' or 'F' or 'Female'}

\item{retireAge}{Retirement age.}

\item{mort.factor}{Mortality factor, default = 1.  This is multiplied by each mortality rate. Values >1 decrease life expectancy.}

\item{mort.adj.years}{Mortality adjustment, default = 0.  This value is added to each generated age at death.}
}
\value{
sim object with person added to simulation
}
\description{
Add Person to persons list in a simulation
}
\examples{
\dontrun{sim1<-addPerson.sim(sim, name, initials, curAge, gender, retireAge, mort.factor, mort.adj.years)}
\dontrun{sim1<-addPerson.sim(sim1,"Rex Macey","RM",56,"M",65,1.0, 0.0)}
}
