% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcAgesForSimulation}
\alias{calcAgesForSimulation}
\title{Calculates ages of persons in Sim}
\usage{
calcAgesForSimulation(sim, asOfDate = Sys.Date())
}
\arguments{
\item{sim}{sim(ulation) object}

\item{asOfDate}{Date to run simulation as of. Used for calculating ages.
default is system date.   May be character YYYY-MM-DD or date.}
}
\value{
sim If there are persons, a curAge is added to each person
}
\description{
Calculates ages of persons in Sim
}
\examples{
\dontrun{calcAgesForSimulation(sim)}
}
