% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcCF}
\alias{calcCF}
\title{Calculate Cash Flow Vector for a trial}
\usage{
calcCF(sim, length, ageDeath1, ageDeath2)
}
\arguments{
\item{sim}{sim(ulation) object}

\item{length}{The number of years in that trial.}

\item{ageDeath1}{age of death of person 1 if applicable}

\item{ageDeath2}{age of death of person 2 if applicable}
}
\value{
vector of cash flows for a single trial
}
\description{
This takes all the cash flows and combines them into a vector of values for a single trial.
Each value in the vector represents a contribution (+) or withdrawal (-) for the  year.
}
\examples{
\dontrun{calcCF(sim,10,82,91)}
}
