% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histReturn_functions.R
\name{calcChronologicalHist}
\alias{calcChronologicalHist}
\title{Calculate Chronological History}
\usage{
calcChronologicalHist(startIdx, n, sbiSub, stockWt)
}
\arguments{
\item{startIdx}{Starting index}

\item{n}{Number of years to generate returns and inflation.}

\item{sbiSub}{A subset of the sbi table limited by sim$minDate and sim$maxDate.}

\item{stockWt}{Weighting to stocks. Bonds will be 1 - StockWt.  Enter 60\% as 0.60.}
}
\value{
data frame with n rows and two columns; return and inflation. These have 1 added to them so 4\% will return as 1.04.
}
\description{
Generates sequential returns and inflation rates using the sbi table.
It will start with the startIdx element of sbiSub and generate n annual
returns using consecutive months.
}
\examples{
\dontrun{calcChronologicalHist(1, 10, sbiSub, 0.6)}
}
