% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcIRR}
\alias{calcIRR}
\title{Calculates the IRR of a stream of cash flow}
\usage{
calcIRR(cashflows)
}
\arguments{
\item{cashflows}{Vector of cash flows. These should include at least one
positive and one negative value.}
}
\value{
Decimal IRR or NA.
}
\description{
Returns the IRR of a stream of cash flows.  NA will be returned
if there isn't at least one positive and one negative value.  If
the IRR is outside the range -99.99\% to 10000\%, NA is likely to be produced.
}
\examples{
\dontrun{calcIRR(cashflows)}
}
