% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcRandReturns}
\alias{calcRandReturns}
\title{Random return(s) from a lognormally distribution in decimal format}
\usage{
calcRandReturns(n, r, sd, t, seed = NA)
}
\arguments{
\item{n}{Number of returns to generate}

\item{r}{Mean return}

\item{sd}{Standard deviation of returns}

\item{t}{Time in years}

\item{seed}{Random seed}
}
\value{
value Random returns in decimal format
}
\description{
Returns n lognormally distributed random variables with a mean of r and a
standard deviation of sd.  Periods over 1 year are annualized.  Specify
inputs in decimal formats (e.g. 8%=0.08 not 8).
}
\examples{
AA_Rand(10,.08,.12,1)

}
