% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_functions.r
\name{calculate_age}
\alias{calculate_age}
\title{Calculate current age}
\usage{
calculate_age(birthdate, asOfDate = Sys.Date())
}
\arguments{
\item{birthdate}{Birthdate in character (YYYY-MM-DD) or as class date.}

\item{asOfDate}{The date to calculate the age. Default is the system date.}
}
\value{
Age in years as numeric
}
\description{
Calculate current age
}
\examples{
\dontrun{calculate_age("1950-01-01")}
}
