% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_functions.r
\name{calculate_age_nearest_birthday}
\alias{calculate_age_nearest_birthday}
\title{Calculate age at nearest birthday}
\usage{
calculate_age_nearest_birthday(birthdate, asOfDate = Sys.Date())
}
\arguments{
\item{birthdate}{Birthdate in character (YYYY-MM-DD) or as class date.}

\item{asOfDate}{The date to calculate the age. Default is the system date.}
}
\value{
Age in years as numeric
}
\description{
Returns the age at the previous or next birthday, whichever is nearest.
}
\examples{
\dontrun{calculate_age_nearest_birthday("1950-01-01")}
}
