% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartDistOfTime}
\alias{chartDistOfTime}
\title{Distribution of Ages/Lengths Chart}
\usage{
chartDistOfTime(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
Chart object.  If nPersons is 0 or 1 there will be one panel representing
the lengths of the trials or the ages of Person 1.  If nPersons is 2, there will
be 3 panels to represent Person 1, Person 2 and the lengths.
}
\description{
This chart shows how long trials last.
}
\examples{
\dontrun{chartDistOfTime(sim)}
}
