% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartDistOfTimeSub}
\alias{chartDistOfTimeSub}
\title{Helper function for the Distribution of Ages/Lengths Chart}
\usage{
chartDistOfTimeSub(rawData, gtitle, xtitle)
}
\arguments{
\item{rawData}{A vector with the lengths of trials or the ages at death}

\item{gtitle}{Main title for the chart}

\item{xtitle}{Title of x-axis}
}
\value{
Chart object. One panel.
}
\description{
Helper function for the Distribution of Ages/Lengths Chart
}
\examples{
\dontrun{chartDistOfTimeSub(rawdata, gtitle, xtitle)}
}
