% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartGrowth1DollarBestMedianWorst}
\alias{chartGrowth1DollarBestMedianWorst}
\title{Chart Growth of $1 for Worst, Median and Best Trials}
\usage{
chartGrowth1DollarBestMedianWorst(sim, simResult)
}
\arguments{
\item{sim}{Simulation object}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
Plotly line chart object
}
\description{
This extracts the returns for the trials with the lowest, median and
highest portfolio values.  It produces a line chart assuming an initial
investment of $1 and no cashflows.  The portfolios values do consider cash
flows.
}
\examples{
\dontrun{chartGrowth1DollarBestMedianWorst(sim, simResult)}
}
