% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartSuccessDonut}
\alias{chartSuccessDonut}
\title{Success Donut Chart}
\usage{
chartSuccessDonut(sim, vs = "T", simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{vs}{Choices are 'T' for target or "Z' for zero}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
a ggplot object
}
\description{
Produces a simple donut chart showing the number and percent of successes
and failures.
}
\examples{
\dontrun{chartSucessDonut(sim)}
}
