% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartSuccessOverTime}
\alias{chartSuccessOverTime}
\title{Success across Time Chart}
\usage{
chartSuccessOverTime(sim, vs = "T", simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{vs}{Choices are 'T' for target or "Z' for zero}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
ggplot2 object
}
\description{
Produces two charts, one over the other.  The top shows stacked bars with
the number of successes and failures by year.  The bottom is a line representing
the cumulative success rate (% of successful trials through that length).  If
the value of vs is T, then the target is used, otherwise zero.
}
\examples{
\dontrun{(sim)}
}
