% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartTerminalValuesVsAverageReturns}
\alias{chartTerminalValuesVsAverageReturns}
\title{Scatterplot of Terminal Values vs Average Returns}
\usage{
chartTerminalValuesVsAverageReturns(sim, simResult, logScale = FALSE)
}
\arguments{
\item{sim}{Simulation object}

\item{simResult}{Name of item in sim with results (class = 'simResult')}

\item{logScale}{TRUE to show the y-axis on a log scale. Default = FALSE.}
}
\value{
Plotly chart object (scatter plot)
}
\description{
Scatterplot of Terminal Values vs Average Returns
}
\examples{
\dontrun{chartTerminalValuesVsAverageReturns(sim, simResult, FALSE)}
}
