% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartValuesOverTime}
\alias{chartValuesOverTime}
\title{Values across Time vs Target}
\usage{
chartValuesOverTime(sim)
}
\arguments{
\item{sim}{Simulation object after running simulate function}
}
\value{
ggplot2 object
}
\description{
Produces a ribbon chart.  There's a line for the median value of all the
trials at each year. Additionally there are 50% and 90% confidence intervals.
A black line shows the value of the target value.
}
\examples{
\dontrun{chartValuesOverTime(sim)}
}
