% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{compareRateOfReturns}
\alias{compareRateOfReturns}
\title{Compare Rates of Return}
\usage{
compareRateOfReturns(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
data frame with the Method, GeomReturn, StdDev, and nObs
}
\description{
Compares the rates of return from different simulations
}
\examples{
\dontrun{compareRateOfReturns(sim)}
}
