% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{findTrialByQuantile}
\alias{findTrialByQuantile}
\title{Find Trial by Quantile}
\usage{
findTrialByQuantile(sim, simResult, qTile)
}
\arguments{
\item{sim}{Simulation object.}

\item{simResult}{Name of item in sim with results (class = 'simResult')}

\item{qTile}{Quantile (0 - 1) of terminal value}
}
\value{
Integer representing which trial terminates with the qTile value.
}
\description{
This function returns the trial number (index) with
a terminal value that matches the quantile (qTile).  So, qTile == 0 will
identify which trial has the lowest value, 1 the highest, and 0.5 the median.
If more than one trial has the same terminal value, the function returns the
trial that has that terminal value earliest.  So, if the quantile of the
terminal values is 0, and there are multiple trials that terminate with 0,
the function returns the trial which ends with 0 earliest.
}
\examples{
\dontrun{findTrialByQuantile(sim, simResult, qTile)}
}
