% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getDatesOfChronologicalTrial}
\alias{getDatesOfChronologicalTrial}
\title{Get Start and End Dates of Chronological Trials By Index}
\usage{
getDatesOfChronologicalTrial(sim, idxVec)
}
\arguments{
\item{sim}{Simulation object}

\item{idxVec}{Vector of indices to a Chronological simulation}
}
\value{
list with startDates and endDates vectors
}
\description{
Given a vector of indices, this will return the starting and ending months of those trials.
So getDatesOfChronologicalTrial(sim, c(3,5)) would return the starting and ending months
for the 3rd and 5th trials.
}
\examples{
\dontrun{getDatesOfChronologicalTrial(sim, idxVec)}
}
