% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getDistOfValuesByYear}
\alias{getDistOfValuesByYear}
\title{Get the Distribution of Value by Year}
\usage{
getDistOfValuesByYear(
  sim,
  probs = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
  simResult = "simulation"
)
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{probs}{numeric vector of probabilities with values [0-1]. The
default is c(0.0, 0.05, 0.25, 0.50, 0.75, 0.95, 1.0).}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
data frame
}
\description{
For each year (1 to maximum length of a trail), it returns the number of
trials of that length (nTrials) and the quantiles associated with the probs.
}
\examples{
\dontrun{getDistOfValuesByYear(sim)}
}
