% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getDistribution}
\alias{getDistribution}
\title{Get the Distribution of a Series}
\usage{
getDistribution(
  x,
  probs = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
  includeGeomAvg = FALSE
)
}
\arguments{
\item{x}{Series for which to get the distribution}

\item{probs}{numeric vector of probabilities with values in [0, 1]}

\item{includeGeomAvg}{boolean to include geometric average}
}
\value{
list with nObs, quanitles associated with the probs, summary, and
the sd (standard deviation)
}
\description{
Intended to return summary statistics for series in the simulation such as
rateOfReturns, agesDeath1 and such.  The geometric average makes sense to
include for rateOfReturn and inflation, but not for ages.  If includeGeomAvg
== TRUE, one will be added to each value for that calculation.
}
\examples{
\dontrun{getDistribution(unlist(sim$simulation$rateOfReturns) - 1,
includeGeomAvg = TRUE)}
}
