% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getGrowth1Dollar}
\alias{getGrowth1Dollar}
\title{Growth of $1}
\usage{
getGrowth1Dollar(sim, simResult, idxVec)
}
\arguments{
\item{sim}{Simulation object.}

\item{simResult}{Name of item in sim with results (class = 'simResult')}

\item{idxVec}{Vector of one or more indices representing a trial.}
}
\value{
Data frame. The number of row will be one more than the longest length of the trials. The
first column is the year and each subsequent column is the growth of $1 for each element in idxVec.
}
\description{
Calculates the growth of $1 from the rates of returns in the
trials specified by idxVec.
}
\examples{
\dontrun{getGrowth1Dollar(sim, simResult, c(1, 2, 3))}
}
