% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getMaxLength}
\alias{getMaxLength}
\title{Get maximum length of all trials}
\usage{
getMaxLength(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
integer
}
\description{
Get maximum length of all trials
}
\examples{
\dontrun{getMaxLength(sim)}
}
