% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getMinLength}
\alias{getMinLength}
\title{Get minimum length of all trials}
\usage{
getMinLength(sim)
}
\arguments{
\item{sim}{Simulation object after running simulate function}
}
\value{
integer
}
\description{
Get minimum length of all trials
}
\examples{
\dontrun{getMinLength(sim)}
}
