% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histReturn_functions.R
\name{getSBIDateRange}
\alias{getSBIDateRange}
\title{Get the Min and Max SBI Dates}
\usage{
getSBIDateRange()
}
\value{
list with minDate and maxDate
}
\description{
Returns a list with the first and last dates of the months in the
Stock, Bonds, Inflation (SBI) table.
}
\examples{
\dontrun{getSBIDateRange()}
}
