% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getSuccessByLength}
\alias{getSuccessByLength}
\title{Get Success Statistics by Length of Trial}
\usage{
getSuccessByLength(sim)
}
\arguments{
\item{sim}{Simulation object after running simulate function}
}
\value{
data frame
}
\description{
Returns a data frame with one row for each length of a trial. It includes
the frequency (number of trials with that length); how many (count) were
succcessful vs the target and vs 0; the percentage success rates, the
cumulative counts and percentages.
}
\examples{
\dontrun{getSuccessByLength(sim)}
}
