% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getSuccessStats}
\alias{getSuccessStats}
\title{Get Success Statistics}
\usage{
getSuccessStats(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
list
}
\description{
Success vs target is defined as a trial with a terminal value >= the
terminal value.   Success vs zero (0) is defined as a trial with a terminal
value > 0. The list returned by this function includes:
vector of success by trial vs target,
vector of success by trial vs 0,
integer count of success vs target,
integet count of success vs 0,
numeric percentage success vs target
numeric percentage success vs 0
}
\examples{
\dontrun{getSuccessStats(sim)}
}
