% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getTerminalValues}
\alias{getTerminalValues}
\title{Get Terminal Values of Each Trial}
\usage{
getTerminalValues(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
vector of values
}
\description{
Returns a vector with a length equal to the number of trials containing the
year by year terminal values. This is useful for determining
success or failure.
}
\examples{
\dontrun{getTerminalValues(sim)}
}
