% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{inflationAdjustPortfolioValues}
\alias{inflationAdjustPortfolioValues}
\title{Inflation Adjust Portfolio Values}
\usage{
inflationAdjustPortfolioValues(sim, simResult)
}
\arguments{
\item{sim}{simulation object}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
List same size as portfolioValues
}
\description{
Inflation Adjust Portfolio Values
}
\examples{
\dontrun{inflationAdjustPortfolioValues(sim, simResult)}
}
