% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{initializeCF}
\alias{initializeCF}
\title{Initialize Cash Flow data frame}
\usage{
initializeCF()
}
\value{
Data frame with no rows but columns for cash flows
}
\description{
Initialize Cash Flow data frame
}
\examples{
\dontrun{initializeCF()}
}
