% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{initializeSim}
\alias{initializeSim}
\title{Intialize a Simulation}
\usage{
initializeSim(
  description,
  nTrials,
  startValue,
  lengthType,
  length = 10,
  seed,
  defaultInflation = 0,
  ror,
  stdDev,
  targetValue = .Machine$double.eps,
  targetValueIsReal = TRUE,
  stockWt = 0.6,
  nConsecMonths = 12,
  retAdj = 0,
  minDate = min(sbi$Month),
  maxDate = max(sbi$Month),
  asOfDate = Sys.Date()
)
}
\arguments{
\item{description}{Description of simulation}

\item{nTrials}{number of trials}

\item{startValue}{Starting dollar value}

\item{lengthType}{'R' for random, or 'F' for fixed. If number of persons is 0, then F is assumed.}

\item{length}{Length for fixed type of simulations in years}

\item{seed}{Seed for random number generator}

\item{defaultInflation}{Inflation rate in decimal}

\item{ror}{Rate of Return in decimal}

\item{stdDev}{Standard Deviation in decimal}

\item{targetValue}{Dollar value used to determine if a trial is a success}

\item{targetValueIsReal}{Logical to indicate if target value should be
adjusted for inflation. Default is TRUE.}

\item{stockWt}{The allocation to stocks, in decimal, used to generate historical random returns.}

\item{nConsecMonths}{Number of consecutive months, 1, 2, 3, 4, 6, or 12 used to generate historical random returns.}

\item{retAdj}{A value to add to each randomly generated annual historic return in decimal.}

\item{minDate}{MinDate Earliest date to use when generating historical random returns.}

\item{maxDate}{MaxDate Latest date to use when generating historical random returns.}

\item{asOfDate}{Date to run the simulation as of. Used when calculating ages.}
}
\value{
List with values describing simulation
}
\description{
Intialize a Simulation
}
\examples{
\dontrun{initializeSim("Sim1 Test", nTrials=500, 1000000, lengthType="R",
length=0, seed=-101, defaultInflation=0, ror=0.10, stdDev=.08,
targetValue=.Machine$double.eps, targetValueIsReal=FALSE)}
}
\seealso{
[calcRandHistReturns()]
}
