% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{initializeSimulation}
\alias{initializeSimulation}
\title{Helper function for Simulation methods}
\usage{
initializeSimulation(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
list with modified sim, sbiSub, and out list
}
\description{
Helper function for Simulation methods
}
\examples{
\dontrun{initializeSimulation(sim)}
}
