% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{makeTrialTable}
\alias{makeTrialTable}
\title{Make a Table Representing a Single Trial}
\usage{
makeTrialTable(sim, simResult, trialNum)
}
\arguments{
\item{sim}{A simulation object}

\item{simResult}{Name of item in sim with results (class = 'simResult')}

\item{trialNum}{Optional - the trial number to create a table for}
}
\value{
a data frame with the Year, PortfolioValue, CashFlow, Return, and Inflation
}
\description{
This function exists to produce a table reflecting the details of a single
trial.   If the trial number is not supplied, a trial is selected with a
length within the 50\% confidence interval and that is closest to the
median terminal value.
}
\examples{
\dontrun{makeTrialTable(sim, simResult, trialNum)}
}
