% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{nCF.sim}
\alias{nCF.sim}
\title{Number of cash flows in a sim(ulation) object}
\usage{
nCF.sim(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
Number of cash flows in object
}
\description{
Number of cash flows in a sim(ulation) object
}
\examples{
\dontrun{nCF.sim(sim1)}
}
