% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{nPersons.sim}
\alias{nPersons.sim}
\title{Number of persons in a sim(ulation) object}
\usage{
nPersons.sim(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
Number of persons in object
}
\description{
Number of persons in a sim(ulation) object
}
\examples{
nPersons.sim(sim1)
}
