% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{print.sim}
\alias{print.sim}
\title{Print Simulation Object}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
\item{x}{Simulation object}

\item{...}{further arguments to or from other methods}
}
\value{
Print result
}
\description{
Print Simulation Object
}
\examples{
\dontrun{print.sim(sim)}
}
