% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{print_sim_cf}
\alias{print_sim_cf}
\title{Print cash flow variables of a simulation object}
\usage{
print_sim_cf(cf)
}
\arguments{
\item{cf}{Cash flow table (data frame)}
}
\value{
Print
}
\description{
Print cash flow variables of a simulation object
}
