% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{print_sim_main}
\alias{print_sim_main}
\title{Print main variables of a simulation object}
\usage{
print_sim_main(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
Print
}
\description{
Print main variables of a simulation object
}
