% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{print_sim_persons}
\alias{print_sim_persons}
\title{Print person variables of a simulation object}
\usage{
print_sim_persons(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
Print
}
\description{
Print person variables of a simulation object
}
