% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{printSub}
\alias{printSub}
\title{Print a line of a simulation object}
\usage{
printSub(desc, value)
}
\arguments{
\item{desc}{Description of what is to be printed}

\item{value}{Formatted value}
}
\value{
Print
}
\description{
Print a line of a simulation object
}
