% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histReturn_functions.R
\name{randHistSub}
\alias{randHistSub}
\title{Subroutine (helper) to calculate random returns using historical data}
\usage{
randHistSub(sbiSub, stockWt, nConsecMonths = 12)
}
\arguments{
\item{sbiSub}{tibble or data frame with historical data}

\item{stockWt}{Weighting to stocks. Bonds will be 1 - StockWt}

\item{nConsecMonths}{Number of consecutive months (default is 12). Must be 1,
2, 3, 4, 6, or 12.}
}
\value{
Two element vector with return and inflation rates + 1
}
\description{
Subroutine (helper) to calculate random returns using historical data
}
\examples{
\dontrun{randHistSub(sbiSub, 0.6)}
}
