% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_Excel.r
\name{readExcelSimData}
\alias{readExcelSimData}
\title{Read Simulation Inputs from Excel}
\usage{
readExcelSimData(xlsxfilename)
}
\arguments{
\item{xlsxfilename}{File name including path}
}
\value{
List with an xl data frame and a xlcf data frame
}
\description{
This functions reads data from the first two tabs of an Excel workbook.
The first is expected to contain general information on the simulation and
the second is expected to contain information on the cash flows.
}
\examples{
\dontrun{readExcelSimData(paste0(testthat::test_path(),"/data/siminput.xlsx"))}
}
