% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_Excel.r
\name{setupSimulation}
\alias{setupSimulation}
\title{Set up (prepare) object for simulation}
\usage{
setupSimulation(simDescription, xl, xlcf)
}
\arguments{
\item{simDescription}{String giving the description of the simulation to prepare.}

\item{xl}{Data frame from readExcelSimData with general simulation inputs.}

\item{xlcf}{Data frame from readExcelSimData with cash flow.}
}
\value{
List with simulation inputs
}
\description{
This can be useful for debugging.
After this function is run, you are ready to step through simulate after this.
}
\examples{
\dontrun{setupSimulation, xl, xlcf}
}
