% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_Excel.r
\name{simulateexcelsub}
\alias{simulateexcelsub}
\title{Sub function for simulating using data from Excel.}
\usage{
simulateexcelsub(simDescription, xl, xlcf)
}
\arguments{
\item{simDescription}{String giving the description of the simulation to prepare.}

\item{xl}{Data frame from readExcelSimData with general simulation inputs.}

\item{xlcf}{Data frame from readExcelSimData with cash flow.}
}
\value{
List with results for the simulation
}
\description{
Sub function for simulating using data from Excel.
}
\examples{
\dontrun{simulateexcelsub("sim1", xl, xlcf)}
}
