% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_Excel.r
\name{simulatefromexcel}
\alias{simulatefromexcel}
\title{Simulate using data from Excel}
\usage{
simulatefromexcel(xlsxfilename)
}
\arguments{
\item{xlsxfilename}{File name including path}
}
\value{
List containing lists for each simulation in the file.
}
\description{
This functions reads data from the first two tabs of an Excel workbook.
The first is expected to contain general information on the simulation and
the second is expected to contain information on the cash flows.
}
\details{
Columns in the first tab might include: description,	nTrials, startValue,
lengthType, length,	seed,	defaultInflation,	ror,	stdDev,	targetValue,
targetValueIsReal,	stockWt,	nConsecMonths,	retAdj,	minDate,	maxDate,
overrideInflation,	asOfDate,	returnGeneratorMethod,	p1name,	p1init,	p1birthdate,
p1gender,	p1retage,	p1mortfactor,	p1mortadjyears,	p2name,	p2init,
p2birthdate,	p2gender,	p2retage,	p2mortfactor,	p2mortadjyears.

Columns in the second tab should include: simulation,	description,
starttype,	start,	endtype,	end,	type,	amount,	defaultInflationAdj,
inflation.
}
\examples{
\dontrun{simulatefromexcel(paste0(testthat::test_path(),"/data/siminput.xlsx"))}
}
