% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{simulateSHC}
\alias{simulateSHC}
\title{Runs simulations on a simulation object}
\usage{
simulateSHC(sim)
}
\arguments{
\item{sim}{sim(ulation) object}
}
\value{
3 lists (simS, simH, and simC) each with 6 items.
  lengths is a vector containing the number of years of each trial.
  agesDeath1 is a vector of the ages of the deaths of person 1.
  agesDeath2 is a vector of the ages of the deaths of person 2.
  cashFlows is a list, each item is a vector of the cashFlows of each trial.
  portfolioValues is a list each item is a vector of portfolio values for each trial.
  ratesOfReturns is a list each item is a vector of 1 + the annual rate of return
}
\description{
Uses the Statistical, Historical and Chronological methods to generate returns.
}
\examples{
\dontrun{simulateSHC(sim)}
}
