% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{simulateStatistical}
\alias{simulateStatistical}
\title{Runs a simulation on a simulation object}
\usage{
simulateStatistical(sim, asOfDate = Sys.Date(), randReturnType = "S")
}
\arguments{
\item{sim}{sim(ulation) object}

\item{asOfDate}{Date to run simulation as of. Used for calculating ages.
default is system date.   May be character YYYY-MM-DD or date.}

\item{randReturnType}{Method used to generate random returns. Either "S" for
statistical or "H" for historical.}
}
\value{
A list with 6 items.
  lengths is a vector containing the number of years of each trial.
  agesDeath1 is a vector of the ages of the deaths of person 1.
  agesDeath2 is a vector of the ages of the deaths of person 2.
  cashFlows is a list, each item is a vector of the cashFlows of each trial.
  portfolioValues is a list each item is a vector of portfolio values for each trial.
  ratesOfReturns is a list each item is a vector of 1 + the annual rate of return
}
\description{
Runs a simulation on a simulation object
}
\examples{
\dontrun{simulateStatistical(sim)}
}
