% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{validateCashFlows}
\alias{validateCashFlows}
\title{Validate Cash Flows}
\usage{
validateCashFlows(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\value{
NULL is all cash flows are valid. Otherwise an error will be generated
}
\description{
Validate Cash Flows
}
\examples{
\dontrun{validateCashFlows(sim)}
}
