% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_classes.r
\name{validateCFType}
\alias{validateCFType}
\title{Validate a Cash Flow Type}
\usage{
validateCFType(strCF)
}
\arguments{
\item{strCF}{String to be tested}
}
\value{
Logical value. TRUE is a valid cash flow type, FALSE if not.
}
\description{
Validate a Cash Flow Type
}
\examples{
validateCFType("p1age")
}
