% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{whichItemsClassName}
\alias{whichItemsClassName}
\title{Identify the Items in a List with a Class Name}
\usage{
whichItemsClassName(sim, nameClass = "simResult")
}
\arguments{
\item{sim}{Simulation object}

\item{nameClass}{name of class}
}
\value{
list with two items. See description.
}
\description{
This will identify all the items in a list that are of a class named
nameClass.  The result is a list with two items.  Indices contains the
indices (locations) of the items, and names contains the name.
}
\examples{
\dontrun{whichItemsClassName(sim, "simResult")}
}
