% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcAgesForSimulation}
\alias{calcAgesForSimulation}
\title{Calculates ages of persons in Sim}
\usage{
calcAgesForSimulation(sim)
}
\arguments{
\item{sim}{sim(ulation) object}
}
\value{
sim If there are persons, a curAge is added to each person
}
\description{
Calculates ages of persons in Sim
}
\examples{
\dontrun{calcAgesForSimulation(sim)}
}
