% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{calcCFSub}
\alias{calcCFSub}
\title{Calculate A Cash Flow Vector}
\usage{
calcCFSub(cf, sim, length, ageDeath1, ageDeath2, histInflation = 0)
}
\arguments{
\item{cf}{description of a cash flow.}

\item{sim}{sim(ulation) object}

\item{length}{The number of years in that trial.}

\item{ageDeath1}{age of death of person 1 if applicable}

\item{ageDeath2}{age of death of person 2 if applicable}

\item{histInflation}{If historical inflation is to be used, this should be
a vector of 1 + inflation rates with length length}
}
\value{
vector of cash flows
}
\description{
This takes a single cash flow description and creates a numeric vector.
These vectors can be combined to create a net flow for each trial.
}
\examples{
\dontrun{calcCF(sim,10,82,91)}
}
