% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartRandomSampleTrialsPortfolioValues}
\alias{chartRandomSampleTrialsPortfolioValues}
\title{Chart a Sample of the Trials (Wealth Over Time)}
\usage{
chartRandomSampleTrialsPortfolioValues(
  sim,
  sampleSize,
  logScale = TRUE,
  simResult = "simulation"
)
}
\arguments{
\item{sim}{Simulation object}

\item{sampleSize}{Size of random sample}

\item{logScale}{True to display the y-axis in log format}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
Chart object (plotly)
}
\description{
This will produce a chart with a sampleSize number of lines representing
individual trials.
}
\examples{
\dontrun{chartRandomSmapleTrialsPortfolioValues(sim, sampleSize, logScale))}
}
