% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartSuccessDonut}
\alias{chartSuccessDonut}
\title{Success Donut Chart}
\usage{
chartSuccessDonut(sim)
}
\arguments{
\item{sim}{Simulation object after running simulate function}
}
\value{
a ggplot object
}
\description{
Produces a simple donut chart showing the number and percent of successes
and failures.
}
\examples{
\dontrun{chartSucessDonut(sim)}
}
