% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSim.r
\name{chartValuesOverTime}
\alias{chartValuesOverTime}
\title{Values across Time vs Target Chart}
\usage{
chartValuesOverTime(
  sim,
  logScale = FALSE,
  simResult = "simulation",
  inflationAdjusted = FALSE
)
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{logScale}{TRUE to display a log scale on the y-axis, FALSE to display a linear scale.}

\item{simResult}{Name of item in sim with results (class = 'simResult')}

\item{inflationAdjusted}{TRUE to adjust portfolio values by inflation, otherwise FALSE.}
}
\value{
ggplot2 object
}
\description{
Produces a ribbon chart.  There's a line for the median value of all the
trials at each year. Additionally there are 50% and 90% confidence intervals.
A black line shows the value of the target value.
}
\examples{
\dontrun{chartValuesOverTime(sim)}
}
