% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasim_functions.r
\name{cvtCF2Yr}
\alias{cvtCF2Yr}
\title{Convert a Cash Flow Type to a Year}
\usage{
cvtCF2Yr(cfType, sim, length, ageDeath1, ageDeath2, value)
}
\arguments{
\item{cfType}{Cash flow type string}

\item{sim}{sim(ulation) object}

\item{length}{length of trial in years}

\item{ageDeath1}{age of death of person 1 if applicable}

\item{ageDeath2}{age of death of person 2 if applicable}

\item{value}{start or end value from cash flow record}
}
\value{
integer representing year in which cash flow starts or ends
}
\description{
cfType must be in "yr,start,end,p1age,p1ret,p1death,p2age,p2ret,p2death,1stdeath,2nddeath"
}
\examples{
cvtCF2Yr(cf$startType,sim,length,ageDeath1,ageDeath2,cf$start)
}
