% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAgesAtDeath.r
\name{generateAgesAtDeath}
\alias{generateAgesAtDeath}
\title{Generate Ages at Death}
\usage{
generateAgesAtDeath(curAge, gender, n = 1, mort.factor = 1)
}
\arguments{
\item{curAge}{Current age of individual.  Assumes just turned this age.}

\item{gender}{'Male' or 'Female'}

\item{n}{Number of ages to generate}

\item{mort.factor}{A value multiplied by each mortality value. Higher than 1 shortens lifespan}
}
\value{
Vector of ages at death
}
\description{
This function generates ages at death depending on a mortality table. If n=1, the function
will return the life expectancy.
}
\examples{
generateAgesAtDeath(51, "Male", 10, 1.0)

}
