% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getDistribution}
\alias{getDistribution}
\title{Get the Distribution of a Series}
\usage{
getDistribution(
  x,
  probs = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
  isReturnSeries = FALSE
)
}
\arguments{
\item{x}{Series for which to get the distribution}

\item{probs}{numeric vector of probabilities with values in [0, 1]}

\item{isReturnSeries}{boolean if a return series. Will include geometric average
and subtract 1 from the probs and summary items.}
}
\value{
list with nObs - the number of observations;
probs - the quanitles associated with the probs;
summary - results of the summary function;
sd - the (standard deviation); and,
geomAvg the geometric average if isReturnSeries is TRUE
}
\description{
Intended to produce summary statistics for series in the simulation such as
rateOfReturns, agesDeath1 and such.  The geometric average makes sense to
include for rateOfReturn and inflation, but not for ages.  If isReturnSeries
== TRUE, the returns and standard deviation will be showin in percent, not
decimal.
}
\examples{
\dontrun{getDistribution(unlist(sim$simulation$rateOfReturns) - 1,
includeGeomAvg = TRUE)}
}
