% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getFrequencyByLength}
\alias{getFrequencyByLength}
\title{Get Frequency of Lengths}
\usage{
getFrequencyByLength(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
data frame
}
\description{
Returns a table in a data frame columns for Lengths and the frequency
(number of trials) of that length.
}
\examples{
\dontrun{getFrequencyByLength(sim)}
}
