% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histReturn_functions.R
\name{getHistoricalReturnStats}
\alias{getHistoricalReturnStats}
\title{Get Historical Return Stats}
\usage{
getHistoricalReturnStats(
  stockWt,
  minDate = min(sbi$Month),
  maxDate = max(sbi$Month)
)
}
\arguments{
\item{stockWt}{Weighting to stocks. Bonds will be 1 - StockWt.  Enter 60\% as 0.60.}

\item{minDate}{Earliest historical date to use. Default is earliest in set.}

\item{maxDate}{Latest historical date to use Default is most recent in set.}
}
\value{
Names numeric vector with StocksGeomRet, BondsGeomRet, PortGeomRet,
StocksSD, BondsSD, PortSD, StocksArithRet, BondsArithRet, PortArithRet in
decimal.
}
\description{
Get Historical Return Stats
}
\examples{
\dontrun{getHistoricalReturnStats(0.6)}
}
