% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getMaxLength}
\alias{getMaxLength}
\title{Get maximum length of all trials}
\usage{
getMaxLength(sim)
}
\arguments{
\item{sim}{Simulation object after running simulate function}
}
\value{
integer
}
\description{
Get maximum length of all trials
}
\examples{
\dontrun{getMaxLength(sim)}
}
