% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getMinLength}
\alias{getMinLength}
\title{Get minimum length of all trials}
\usage{
getMinLength(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
integer
}
\description{
Get minimum length of all trials
}
\examples{
\dontrun{getMinLength(sim)}
}
