% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSim.r
\name{getTargetValues}
\alias{getTargetValues}
\title{Get Target Values}
\usage{
getTargetValues(sim, simResult = "simulation")
}
\arguments{
\item{sim}{Simulation object after running simulate function}

\item{simResult}{Name of item in sim with results (class = 'simResult')}
}
\value{
vector of values
}
\description{
Returns a vector of length = longest trial + 1. It contain the
year by year target values. This is useful for comparing ending values of
trials to determine success or failure.
}
\examples{
\dontrun{getTargetValues(sim)}
}
